refactor(helpers): create helpers folder
