+++ src/helpers/__tests__/highlight-test.js
