+++ src/helpers/index.js
@@ -1,0 +1,1 @@
+export { default as highlight } from './highlight.js';
