+++ src/lib/main.js
@@ -5,10 +5,10 @@ import algoliasearchHelper from 'algoliasearch-helper';
 
 import InstantSearch from './InstantSearch.js';
 import version from './version.js';
-import highlight from './highlight.js';
 
 import * as connectors from '../connectors/index.js';
 import * as widgets from '../widgets/index.js';
+import * as helpers from '../helpers/index.js';
 
 import * as routers from './routers/index.js';
 import * as stateMappings from './stateMappings/index.js';
@@ -180,6 +180,6 @@ instantsearch.createQueryString =
 instantsearch.connectors = connectors;
 instantsearch.widgets = widgets;
 instantsearch.version = version;
-instantsearch.highlight = highlight;
+instantsearch.highlight = helpers.highlight;
 
 export default instantsearch;
